// ODOT.ARMS.Web/ClientApp/src/app/project/models/cb-raw.ts
export class ControllingBoardRaw {
  controllingBoardId: string;
  projectId: string;
  controllingBoardNumber: string;
  controllingBoardType: number;
  controllingBoardStatus: number;
  publicCommentText: string;
  userId: string;
  entryDate: Date;
  activeInd: string;
  document: string;
  controllingBoardDate: Date;
  submissionDate: Date;
  controllingBoardCategory: number;
  docCnt: number;
}